Refactor Nacos configuration and registration logic; move initConfig to client.go and create separate register function in naming.go
