+++ internal/infrastructure/nacos/client.go
@@ -4,13 +4,43 @@ import (
 	"github.com/nacos-group/nacos-sdk-go/v2/clients"
 	"github.com/nacos-group/nacos-sdk-go/v2/clients/config_client"
 	"github.com/nacos-group/nacos-sdk-go/v2/clients/naming_client"
+	"github.com/nacos-group/nacos-sdk-go/v2/common/constant"
 	"github.com/nacos-group/nacos-sdk-go/v2/vo"
+	"nekoacm-server/pkg/config"
 )
 
 var (
-	namingClient naming_client.INamingClient
-	configClient config_client.IConfigClient
+	clientConfig  constant.ClientConfig
+	serverConfigs []constant.ServerConfig
+	namingClient  naming_client.INamingClient
+	configClient  config_client.IConfigClient
 )
+
+func initConfig() {
+	cConf := config.Conf.Nacos.Client
+	sConfs := config.Conf.Nacos.Server
+
+	// 创建clientConfig
+	clientConfig = *constant.NewClientConfig(
+		constant.WithNamespaceId(cConf.NamespaceId),
+		constant.WithTimeoutMs(cConf.TimeoutMs),
+		constant.WithNotLoadCacheAtStart(cConf.NotLoadCacheAtStart),
+		constant.WithLogDir(cConf.LogDir),
+		constant.WithCacheDir(cConf.CacheDir),
+		constant.WithLogLevel(cConf.LogLevel),
+	)
+
+	// 创建serverConfig
+	serverConfigs = make([]constant.ServerConfig, 0, len(sConfs))
+	for _, conf := range sConfs {
+		serverConfigs = append(serverConfigs, *constant.NewServerConfig(
+			conf.IpAddr,
+			conf.Port,
+			constant.WithScheme(conf.Scheme),
+			constant.WithContextPath(conf.ContextPath),
+		))
+	}
+}
 
 func initClient() error {
 	var err error
